+++ MyDAL/Interfaces/ICount.cs
@@ -7,13 +7,13 @@ namespace MyDAL.Interfaces
     internal interface ICount<M>
         where M:class
     {
-        Task<long> CountAsync();
-        Task<long> CountAsync<F>(Expression<Func<M, F>> func);
+        Task<int> CountAsync();
+        Task<int> CountAsync<F>(Expression<Func<M, F>> func);
     }
 
     internal interface ICountX
     {
-        Task<long> CountAsync();
-        Task<long> CountAsync<F>(Expression<Func<F>> func);
+        Task<int> CountAsync();
+        Task<int> CountAsync<F>(Expression<Func<F>> func);
     }
 }
